restructure file for test
