+++ src/Interface.js
@@ -1,10 +1,12 @@
 import Domelements from './Domelements';
+import Navbar from './Navbar';
 
 const Dom = Domelements();
+const Mynav=Navbar();
 
 const Interface = () => {
   const container = document.querySelector('.container');
-  container.appendChild(Dom.navgenerator());
+  container.appendChild(Mynav.navgenerator());
   container.appendChild(Dom.taskinfo());
   container.appendChild(Dom.bodycont());
 };
